requirements.txt
aiofiles==0.8.0
anyio==3.6.1
appdirs==1.4.4
astroid==2.12.10
attrs==21.4.0
brotlipy==0.7.0
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cmake==3.22.5
colorama==0.4.5
croniter==1.3.7
Cython==0.29.30
diff-cover==7.0.1
dill==0.3.5.1
fastapi~=0.85.0
fastapi-pagination==0.9.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.5.0
iniconfig==1.1.1
isort==5.10.1
Jinja2~=3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
minio==6.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.29
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
pure-sasl==0.6.2
py==1.11.0
pydantic~=1.10.2
PyHive==0.6.5
pylint==2.15.3
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
sasl==0.3.1
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy~=1.4.39
sqlfluff==1.3.2
starlette==0.19.1
tblib==1.7.0
thrift==0.16.0
thrift-sasl==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
tqdm==4.64.1
typing_extensions==4.3.0
uvicorn~=0.18.3
uvloop==0.17.0
wrapt==1.14.1

python-kubernetes
decorator~=5.1.1
configparser~=5.3.0